+++ client/views/index.js
@@ -32,3 +32,5 @@ export { default as text } from './text.js';
 export { default as textMatch } from './text-match.js';
 export { default as tocItem } from './toc-item.js';
 export { default as tocSection } from './toc-section.js';
+export { default as treeItem } from './tree-leaf.js';
+export { default as tree } from './tree.js';