+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## next
+
+- Client
+    - Added context menu in `struct` view with `Copy as JSON` actions
+    - Improved copying performance of a large content to the buffer
+    - Fixed report content hide on query error
+
 ## 1.0.0-beta.6 (16-01-2019)
 
 - Build